Wrap the amplitude up in a class

+++ test/test_visualisation.py
@@ -2,9 +2,10 @@ from test.configurable import Configurable
 
 from impact.datafit import DataFit
 from impact.datapoint import DataSet
+from impact.parametrization.numeric import Numeric
 
 class TestVisualRepresentation(Configurable):
 
     def testValues(self):
-        df = DataFit('a')
+        df = DataFit('a', Numeric())
         df.fit(self.dataset)
